+++ annotations/nocedit.exe/pseudocode/src/core/game.cpp/CGame_saveGame_FUN_004e0cd0.cpp
@@ -162,8 +162,7 @@ void __cdecl core_game_cpp_CGame_saveGame_FUN_004e0cd0(CGame *this_ptr,char *sav
     crt_fstream_cpp_ostream_write_FUN_005ffcb3(&local_88,"EFD",4);
     support_codec_cpp_CLZWCompress_ctor_FUN_0043f2d0(&local_50,0x8000,8);
     support_codec_cpp_CLZWCompress_init_FUN_0043f320(&local_50);
-    support_codec_cpp_CLZWCompress_process_FUN_0043f360
-              (&local_50,&local_108,(int)&local_18,&local_88);
+    support_codec_cpp_CLZWCompress_process_FUN_0043f360(&local_50,&local_108,&local_18,&local_88);
     support_codec_cpp_CLZWCompress_finalize_FUN_0043f440(&local_50,&local_88);
     _fflush(local_14c);
     _fflush(local_cc);
